+++ clients/ctl/admin-ui/src/features/data-subjects/data-subject.slice.ts
@@ -37,11 +37,31 @@ export const dataSubjectsApi = createApi({
       }),
       invalidatesTags: ["Data Subjects"],
     }),
+    createDataSubject: build.mutation<DataSubject, DataSubject>({
+      query: (dataSubject) => ({
+        url: `data_subject/`,
+        method: "POST",
+        body: dataSubject,
+      }),
+      invalidatesTags: ["Data Subjects"],
+    }),
+    deleteDataSubject: build.mutation<string, string>({
+      query: (key) => ({
+        url: `data_subject/${key}`,
+        params: { resource_type: "data_subject" },
+        method: "DELETE",
+      }),
+      invalidatesTags: ["Data Subjects"],
+    }),
   }),
 });
 
-export const { useGetAllDataSubjectsQuery, useUpdateDataSubjectMutation } =
-  dataSubjectsApi;
+export const {
+  useGetAllDataSubjectsQuery,
+  useUpdateDataSubjectMutation,
+  useCreateDataSubjectMutation,
+  useDeleteDataSubjectMutation,
+} = dataSubjectsApi;
 
 export const dataSubjectsSlice = createSlice({
   name: "dataSubjects",