+++ clients/ctl/admin-ui/src/features/taxonomy/TaxonomyTabContent.tsx
@@ -1,23 +1,65 @@
-import { Center, SimpleGrid, Spinner, Text } from "@fidesui/react";
-import { useMemo, useState } from "react";
+import {
+  Center,
+  SimpleGrid,
+  Spinner,
+  Tag,
+  Text,
+  useDisclosure,
+  useToast,
+} from "@fidesui/react";
+import { useEffect, useMemo, useState } from "react";
 
-import AccordionTree from "../common/AccordionTree";
+import { useAppDispatch, useAppSelector } from "~/app/hooks";
+import AccordionTree from "~/features/common/AccordionTree";
+import ConfirmationModal from "~/features/common/ConfirmationModal";
+import { getErrorMessage } from "~/features/common/helpers";
+import { errorToastParams, successToastParams } from "~/features/common/toast";
+import { TreeNode } from "~/features/common/types";
+import { isErrorResult } from "~/types/errors";
+
+import ActionButtons from "./ActionButtons";
 import { transformTaxonomyEntityToNodes } from "./helpers";
 import { TaxonomyHookData } from "./hooks";
+import { selectIsAddFormOpen, setIsAddFormOpen } from "./taxonomy.slice";
 import TaxonomyFormBase from "./TaxonomyFormBase";
 import { TaxonomyEntity, TaxonomyEntityNode } from "./types";
+
+const CustomTag = ({ node }: { node: TaxonomyEntityNode }) => {
+  const { is_default: isDefault } = node;
+  return !isDefault ? (
+    <Tag
+      backgroundColor="gray.500"
+      color="white"
+      size="sm"
+      height="fit-content"
+      data-testid={`custom-tag-${node.label}`}
+    >
+      Custom
+    </Tag>
+  ) : null;
+};
 
 interface Props {
   useTaxonomy: () => TaxonomyHookData<TaxonomyEntity>;
 }
 
+const DEFAULT_INITIAL_VALUES: TaxonomyEntity = {
+  fides_key: "",
+  parent_key: "",
+  name: "",
+  description: "",
+};
+
 const TaxonomyTabContent = ({ useTaxonomy }: Props) => {
+  const dispatch = useAppDispatch();
   const {
     isLoading,
     data,
     labels,
-    edit: onEdit,
-    extraFormFields,
+    handleCreate: createEntity,
+    handleEdit,
+    handleDelete: deleteEntity,
+    renderExtraFormFields,
     transformEntityToInitialValues,
   } = useTaxonomy();
   const taxonomyNodes = useMemo(() => {
@@ -28,6 +70,28 @@ const TaxonomyTabContent = ({ useTaxonomy }: Props) => {
   }, [data]);
 
   const [editEntity, setEditEntity] = useState<TaxonomyEntity | null>(null);
+
+  const isAdding = useAppSelector(selectIsAddFormOpen);
+
+  useEffect(() => {
+    // prevent both the add and edit forms being opened at once
+    if (isAdding) {
+      setEditEntity(null);
+    }
+  }, [isAdding]);
+
+  const closeAddForm = () => {
+    dispatch(setIsAddFormOpen(false));
+  };
+
+  const [deleteKey, setDeleteKey] = useState<string | null>(null);
+
+  const {
+    isOpen: deleteIsOpen,
+    onOpen: onDeleteOpen,
+    onClose: onDeleteClose,
+  } = useDisclosure();
+  const toast = useToast();
 
   if (isLoading) {
     return (
@@ -40,29 +104,86 @@ const TaxonomyTabContent = ({ useTaxonomy }: Props) => {
     return <Text>Could not find data.</Text>;
   }
 
-  const handleSetEditEntity = (node: TaxonomyEntityNode) => {
+  const taxonomyType = labels.fides_key.toLocaleLowerCase();
+
+  const handleSetEditEntity = (node: TreeNode) => {
+    if (isAdding) {
+      closeAddForm();
+    }
     const entity = data?.find((d) => d.fides_key === node.value) ?? null;
     setEditEntity(entity);
   };
 
+  const handleSetDeleteKey = (node: TreeNode) => {
+    setDeleteKey(node.value);
+    onDeleteOpen();
+  };
+
+  const handleDelete = async () => {
+    if (deleteKey) {
+      const result = await deleteEntity(deleteKey);
+      if (isErrorResult(result)) {
+        toast(errorToastParams(getErrorMessage(result.error)));
+      } else {
+        toast(successToastParams(`Successfully deleted ${taxonomyType}`));
+      }
+      onDeleteClose();
+      setEditEntity(null);
+    }
+  };
+
   return (
-    <SimpleGrid columns={2} spacing={2}>
-      <AccordionTree
-        nodes={taxonomyNodes}
-        onEdit={handleSetEditEntity}
-        focusedKey={editEntity?.fides_key}
-      />
-      {editEntity ? (
-        <TaxonomyFormBase
-          labels={labels}
-          entity={editEntity}
-          onCancel={() => setEditEntity(null)}
-          onEdit={onEdit}
-          extraFormFields={extraFormFields}
-          initialValues={transformEntityToInitialValues(editEntity)}
+    <>
+      <SimpleGrid columns={2} spacing={2}>
+        <AccordionTree
+          nodes={taxonomyNodes}
+          focusedKey={editEntity?.fides_key}
+          renderHover={(node) => (
+            <ActionButtons
+              onDelete={handleSetDeleteKey}
+              onEdit={handleSetEditEntity}
+              node={node as TaxonomyEntityNode}
+            />
+          )}
+          renderTag={(node) => <CustomTag node={node as TaxonomyEntityNode} />}
         />
-      ) : null}
-    </SimpleGrid>
+        {editEntity ? (
+          <TaxonomyFormBase
+            labels={labels}
+            onCancel={() => setEditEntity(null)}
+            onSubmit={handleEdit}
+            renderExtraFormFields={renderExtraFormFields}
+            initialValues={transformEntityToInitialValues(editEntity)}
+          />
+        ) : null}
+        {isAdding ? (
+          <TaxonomyFormBase
+            labels={labels}
+            onCancel={closeAddForm}
+            onSubmit={createEntity}
+            renderExtraFormFields={renderExtraFormFields}
+            initialValues={transformEntityToInitialValues(
+              DEFAULT_INITIAL_VALUES
+            )}
+          />
+        ) : null}
+      </SimpleGrid>
+      <ConfirmationModal
+        isOpen={deleteIsOpen}
+        onClose={onDeleteClose}
+        onConfirm={handleDelete}
+        title={`Delete ${taxonomyType}`}
+        message={
+          <Text>
+            You are about to permanently delete the {taxonomyType}{" "}
+            <Text color="complimentary.500" as="span" fontWeight="bold">
+              {deleteKey}
+            </Text>{" "}
+            from your taxonomy. Are you sure you would like to continue?
+          </Text>
+        }
+      />
+    </>
   );
 };