+++ src/fides/ctl/core/config/user_settings.py
@@ -4,7 +4,7 @@
 
 from typing import Dict, Optional
 
-from pydantic import BaseModel, validator
+from pydantic import validator
 
 from fides.ctl.core.utils import generate_request_headers
 
@@ -24,9 +24,8 @@ class UserSettings(FidesSettings):
 
     # Automatically generate the request_headers on object creation
     @validator("request_headers", pre=True, always=True)
-    def get_request_headers(
-        cls: BaseModel, value: Optional[Dict], values: Dict
-    ) -> Dict[str, str]:
+    @classmethod
+    def get_request_headers(cls, value: Optional[Dict], values: Dict) -> Dict[str, str]:
         return generate_request_headers(values["user_id"], values["api_key"])
 
     class Config: